requirements.txt
# Основные зависимости для RAG-ассистента

# OpenAI API клиент для работы с LLM и эмбеддингами
openai>=1.12.0

# FAISS для локального векторного хранилища (совместим с Python 3.14)
faiss-cpu>=1.7.4

# Загрузка переменных окружения из .env файла
python-dotenv>=1.0.0

